01-docker-terraform/2_docker_sql/requirements.txt
pandas
sqlalchemy
psycopg2-binary
pyarrow
wget
requests
pyyaml